requirements.txt
numpy
scikit-learn
scikit-image
matplotlib
# pytorch
torchsummary
livelossplot
pytest
tqdm
IPython
flake8
